requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.28.1
colorama==0.4.6
comm==0.2.1
compose==1.5.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
gast==0.5.4
gdown==4.7.1
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.22.0
gradio-client==0.13.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.2
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.0
jupyter-core==5.7.1
jupyterlab-widgets==3.0.10
keras==2.15.0
kiwisolver==1.4.5
kornia==0.7.1
lazy-loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.1
numba==0.58.1
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.15
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
pillow-heif==0.15.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycolmap==0.4.0
pydantic==2.6.4
pydantic-core==2.16.3
pydub==0.25.1
pyglet==1.5.28
pygments==2.17.2
pyparsing==3.1.1
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
roma==1.4.3
rpds-py==0.17.1
rsa==4.9
ruff==0.3.3
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.15.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torch-dimcheck
torch-localize
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
trimesh==4.2.0
triton==2.2.0
typer==0.9.0
typing-extensions==4.10.0
tzdata==2023.4
urllib3==1.26.13
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.14.1

